Convert to long

+++ src/executor.h
@@ -1,4 +1,4 @@
 #include "parser.h"
 
 void executor (ast_node *);
-int execute_node(ast_node *);
+long execute_node(ast_node *);
